BackPython/requirements.txt
attrs==22.2.0
bidict==0.21.4
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.0.4
clickclick==20.10.2
colorama==0.4.5
connexion==2.14.2
dataclasses==0.8
dnspython==2.2.1
eventlet==0.33.3
Flask==2.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
greenlet==2.0.2
h11==0.13.0
idna==3.10
importlib-metadata==4.8.3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0
jsonschema==4.0.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.29
packaging==21.3
protobuf==3.19.6
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.6.0
python-engineio==4.10.1
python-socketio==5.11.0
PyYAML==6.0.1
requests==2.27.1
simple-websocket==1.1.0
six==1.17.0
swagger-ui-bundle==0.0.9
typing-extensions==4.1.1
urllib3==1.26.20
Werkzeug==2.0.3
wsproto==0.14.1
zipp==3.6.0
